Create README.md
Mining cryptocurrencies such as Ethereum and Bitcoin involves completing a proof-of-work where miners solve complex puzzles to earn profit in terms of the cryptocurrency they mined.

Owing to the complexity of the puzzles, mining cryptocurrency often comes with a high computational cost. Developers of large cryptocurrency mining operations have created workarounds that make use of collective resources from users to complete the calculation needed for a proof-of-work.

Cryptojacking is a type of cyberattack that uses malicious scripts similar to those from large cryptocurrency houses to illegally mine cryptocurrency on other people’s machines without their users being aware. These attacks are stealthy and difficult to detect or analyse, often leading to decreased computing speeds for users, as well as crashes due to the strain on computational resources.